packages/sdk_python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "context_protocol_sdk"
version = "0.1.0"
authors = [
  { name="Context Protocol Team" },
]
description = "Python SDK for Context Protocol"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "web3>=6.0.0",
    "python-dotenv>=1.0.0",
    "eth-typing>=3.0.0",
    "eth-utils>=2.0.0"
]

[project.urls]
"Homepage" = "https://github.com/contextprotocol/core"
